% Generated by tinyroxygen: do not edit by hand
% Please edit documentation in cpp4rsimplexphase2-package.R
\name{cpp4r_simplex_phase2}
\alias{cpp4r_simplex_phase2}
\title{Print the table of the simplex algorithm}
\usage{
cpp4r_simplex_phase2(c, b, A)
}

\arguments{
\item{c}{vector of coefficients of the objective function}

\item{b}{vector of the right hand side of the constraints}

\item{A}{matrix of the coefficients of the constraints}
}

\description{
Print the table of the simplex algorithm
}

