% Generated by tinyroxygen: do not edit by hand
% Please edit documentation in cpp4rsimplexphase2-package.R
\name{cpp4rsimplexphase2-package}
\alias{cpp4rsimplexphase2-package}
\title{A Naive Implementation of the Simplex Phase 2 Algorithm}
\description{
Just to demonstrate how to print outputs and do some simple linear
 algebra. This is an extremely naive implementation.
}

